Created a functional game with one script.
More or less recreated the entire game/project in a single script.
New single-script game has:
- Game Engine Functionality by attaching scripts and other components (such as Audio Source) to game objects. Debug functionality and input response (limited to spacebar).
- Appropriate naming conventions. Pascal Casing for Methods and Classes, camelCasing for variables.
- Comments on code, commented out.
- Debugging of code. Math equations in debug log. Display of variables and context.
- Creating variables and assigning values. Variables created before Start(), created after start so that variables always start with the same value.
Assigning and using random values. Attack and defence multipliers are used to vary chances of success/failure in combat. They are also used with health regeneration and experience gain.
If statements and conditional logic. A whoooole lot of it.
Using operands - all of the combat is based on operands.
Basic win/loss condition that ends/restarts the game. Created methods for win AND lose conditions (win is level == 5, lose is playerHealth <= 0). Struggling to get a proper loop, have to exit game to replay.